feat: Add BookControlMenu
Added book control menu to allow for moving a book between bookshelves. Also added BookshelvesContext component to avoid prop drilling to the menu since it needs the bookshelves to build the menu list dynamically and needed a handler to update the books API and local state.

Double-checked App.js variables to ensure they're not rendered unnecessarily.

Added a loading and error state in App.js with CSS in App.css.